[EDELIVERY-11169] Improve code coverage fix test

+++ Core/Domibus-MSH/src/main/java/eu/domibus/core/message/retention/MessageRetentionDefaultService.java
@@ -99,7 +99,9 @@ public class MessageRetentionDefaultService implements MessageRetentionService {
                 LOG.warn("MessageId [{}] not found", messageId);
             }
         }
-        userMessageDefaultService.deleteMessages(allMessages);
+        if (allMessages.size() > 0) {
+            userMessageDefaultService.deleteMessages(allMessages);
+        }
     }
 
     /**
